# internal/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Griyo HTTP API",
    "version": "0.1.0",
    "description": "Owner cookie session; runner bearer token. gRPC contracts are in proto/griyo.proto."
  },
  "paths": {
...
    "/api/approvals": {
      "get": {
        "summary": "List approvals",
        "security": [
          {
            "ownerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Approval"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Login required"
          },
          "409": {
            "description": "Invalid state"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "summary": "Owner login",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Login required"
          },
          "409": {
            "description": "Invalid state"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/session": {
      "get": {
        "summary": "Current session",
        "security": [
          {
            "ownerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Login required"
          },
          "409": {
            "description": "Invalid state"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Log out",
        "security": [
          {
            "ownerSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Login required"
          },
          "409": {
            "description": "Invalid state"
          }
        }
      }
    },
...
    "/api/runner/heartbeat": {
      "post": {
        "summary": "Runner heartbeat",
        "security": [
          {
            "runnerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Server"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Login required"
          },
          "409": {
            "description": "Invalid state"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Metrics"
              }
            }
          }
        }
      }
    },
    "/api/runner/claim": {
      "post": {
        "summary": "Runner claim",
        "security": [
          {
            "runnerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Login required"
          },
          "409": {
            "description": "Invalid state"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServerConfig": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "fallbackModel": {
            "type": "string"
          },
          "credentialRef": {
            "type": "string"
          },
          "modelOverrides": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "dailyBudget": {
            "type": "number"
          },
          "maxConcurrency": {
            "type": "integer"
          },
          "timeoutMinutes": {
            "type": "integer"
          },
          "maxRetries": {
            "type": "integer"
          },
          "requireApproval": {
            "type": "boolean"
          },
          "isolatedWorkspace": {
            "type": "boolean"
          },
          "stopOnBudget": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "Server": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "config": {
            "$ref": "#/components/schemas/ServerConfig"
          },
          "metrics": {
            "$ref": "#/components/schemas/Metrics"
          },
          "lastSeen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Metrics": {
        "type": "object",
        "properties": {
          "cpuPercent": {
            "type": "number"
          },
          "memoryUsedGb": {
            "type": "number"
          },
          "memoryTotalGb": {
            "type": "number"
          },
          "diskUsedGb": {
            "type": "number"
          },
          "diskTotalGb": {
            "type": "number"
          },
          "containers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Container"
            }
          }
        },
        "additionalProperties": false
      },
      "Container": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "brief": {
            "type": "string"
          },
          "criteria": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "agentId": {
            "type": "string"
          },
          "workServerId": {
            "type": "string"
          },
          "deployServerId": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "budget": {
            "type": "number"
          },
          "progress": {
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "result": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "branch": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "spent": {
            "type": "number"
          },
          "leaseId": {
            "type": "string"
          },
          "execution": {
            "$ref": "#/components/schemas/TaskExecution"
          }
        },
        "additionalProperties": false
      },
      "TaskExecution": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "credentialRef": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "repositoryUrl": {
            "type": "string"
          },
          "timeoutMinutes": {
            "type": "integer"
          },
          "maxRetries": {
            "type": "integer"
          }
        },
        "additionalProperties": false
      },
...
      "Approval": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "serverId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ownerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "griyoSession"
      },
      "runnerToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}